Refactor, merge SilverTextDocumentService and SilverWorkspaceService
